feat: Move TrivialApp tests into one file ForceMoveApp test was actually testing TrivialApp. This moves that test into the new TrivialApp test file
